remove transient IDbConnection (#678)

+++ ErsatzTV.Infrastructure/Data/Repositories/LibraryRepository.cs
@@ -1,5 +1,4 @@
-﻿using System.Data;
-using Dapper;
+﻿using Dapper;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Interfaces.Metadata;
 using ErsatzTV.Core.Interfaces.Repositories;
@@ -9,32 +8,27 @@ namespace ErsatzTV.Infrastructure.Data.Repositories;
 
 public class LibraryRepository : ILibraryRepository
 {
-    private readonly IDbConnection _dbConnection;
     private readonly ILocalFileSystem _localFileSystem;
     private readonly IDbContextFactory<TvContext> _dbContextFactory;
 
-    public LibraryRepository(
-        ILocalFileSystem localFileSystem,
-        IDbContextFactory<TvContext> dbContextFactory,
-        IDbConnection dbConnection)
+    public LibraryRepository(ILocalFileSystem localFileSystem, IDbContextFactory<TvContext> dbContextFactory)
     {
         _localFileSystem = localFileSystem;
         _dbContextFactory = dbContextFactory;
-        _dbConnection = dbConnection;
     }
 
     public async Task<LibraryPath> Add(LibraryPath libraryPath)
     {
-        await using TvContext context = _dbContextFactory.CreateDbContext();
-        await context.LibraryPaths.AddAsync(libraryPath);
-        await context.SaveChangesAsync();
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        await dbContext.LibraryPaths.AddAsync(libraryPath);
+        await dbContext.SaveChangesAsync();
         return libraryPath;
     }
 
-    public Task<Option<Library>> Get(int libraryId)
+    public async Task<Option<Library>> Get(int libraryId)
     {
-        using TvContext context = _dbContextFactory.CreateDbContext();
-        return context.Libraries
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Libraries
             .Include(l => l.Paths)
             .ThenInclude(p => p.LibraryFolders)
             .OrderBy(l => l.Id)
@@ -42,37 +36,45 @@ public class LibraryRepository : ILibraryRepository
             .Map(Optional);
     }
 
-    public Task<Option<LocalLibrary>> GetLocal(int libraryId)
+    public async Task<Option<LocalLibrary>> GetLocal(int libraryId)
     {
-        using TvContext context = _dbContextFactory.CreateDbContext();
-        return context.LocalLibraries
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.LocalLibraries
             .OrderBy(l => l.Id)
             .SingleOrDefaultAsync(l => l.Id == libraryId)
             .Map(Optional);
     }
 
-    public Task<List<Library>> GetAll()
+    public async Task<List<Library>> GetAll()
     {
-        using TvContext context = _dbContextFactory.CreateDbContext();
-        return context.Libraries
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Libraries
             .AsNoTracking()
             .Include(l => l.MediaSource)
             .Include(l => l.Paths)
             .ToListAsync();
     }
 
-    public Task<Unit> UpdateLastScan(Library library) => _dbConnection.ExecuteAsync(
-        "UPDATE Library SET LastScan = @LastScan WHERE Id = @Id",
-        new { library.LastScan, library.Id }).ToUnit();
-
-    public Task<Unit> UpdateLastScan(LibraryPath libraryPath) => _dbConnection.ExecuteAsync(
-        "UPDATE LibraryPath SET LastScan = @LastScan WHERE Id = @Id",
-        new { libraryPath.LastScan, libraryPath.Id }).ToUnit();
-
-    public Task<List<LibraryPath>> GetLocalPaths(int libraryId)
+    public async Task<Unit> UpdateLastScan(Library library)
     {
-        using TvContext context = _dbContextFactory.CreateDbContext();
-        return context.LocalLibraries
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Connection.ExecuteAsync(
+            "UPDATE Library SET LastScan = @LastScan WHERE Id = @Id",
+            new { library.LastScan, library.Id }).ToUnit();
+    }
+
+    public async Task<Unit> UpdateLastScan(LibraryPath libraryPath)
+    {
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Connection.ExecuteAsync(
+            "UPDATE LibraryPath SET LastScan = @LastScan WHERE Id = @Id",
+            new { libraryPath.LastScan, libraryPath.Id }).ToUnit();
+    }
+
+    public async Task<List<LibraryPath>> GetLocalPaths(int libraryId)
+    {
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.LocalLibraries
             .Include(l => l.Paths)
             .OrderBy(l => l.Id)
             .SingleOrDefaultAsync(l => l.Id == libraryId)
@@ -80,62 +82,76 @@ public class LibraryRepository : ILibraryRepository
             .Match(l => l.Paths, () => new List<LibraryPath>());
     }
 
-    public Task<Option<LibraryPath>> GetPath(int libraryPathId)
+    public async Task<Option<LibraryPath>> GetPath(int libraryPathId)
     {
-        using TvContext context = _dbContextFactory.CreateDbContext();
-        return context.LibraryPaths
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.LibraryPaths
             .OrderBy(lp => lp.Id)
             .SingleOrDefaultAsync(lp => lp.Id == libraryPathId)
             .Map(Optional);
     }
 
-    public Task<int> CountMediaItemsByPath(int libraryPathId) =>
-        _dbConnection.QuerySingleAsync<int>(
+    public async Task<int> CountMediaItemsByPath(int libraryPathId)
+    {
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Connection.QuerySingleAsync<int>(
             @"SELECT COUNT(*) FROM MediaItem WHERE LibraryPathId = @LibraryPathId",
             new { LibraryPathId = libraryPathId });
+    }
 
-    public Task<List<int>> GetMediaIdsByLocalPath(int libraryPathId) =>
-        _dbConnection.QueryAsync<int>(
+    public async Task<List<int>> GetMediaIdsByLocalPath(int libraryPathId)
+    {
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+        return await dbContext.Connection.QueryAsync<int>(
                 @"SELECT Id FROM MediaItem WHERE LibraryPathId = @LibraryPathId",
                 new { LibraryPathId = libraryPathId })
             .Map(result => result.ToList());
+    }
 
     public async Task DeleteLocalPath(int libraryPathId)
     {
-        await using TvContext context = _dbContextFactory.CreateDbContext();
+        await using TvContext context = await _dbContextFactory.CreateDbContextAsync();
         LibraryPath libraryPath = await context.LibraryPaths.FindAsync(libraryPathId);
-        context.LibraryPaths.Remove(libraryPath);
-        await context.SaveChangesAsync();
+        if (libraryPath != null)
+        {
+            context.LibraryPaths.Remove(libraryPath);
+            await context.SaveChangesAsync();
+        }
     }
 
-    public Task<Unit> SetEtag(
+    public async Task<Unit> SetEtag(
         LibraryPath libraryPath,
         Option<LibraryFolder> knownFolder,
         string path,
-        string etag) =>
-        knownFolder.Match(
+        string etag)
+    {
+        return await knownFolder.Match(
             async folder =>
             {
-                await _dbConnection.ExecuteAsync(
+                await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+                await dbContext.Connection.ExecuteAsync(
                     "UPDATE LibraryFolder SET Etag = @Etag WHERE Id = @Id",
                     new { folder.Id, Etag = etag });
             },
             async () =>
             {
-                await using TvContext context = await _dbContextFactory.CreateDbContextAsync();
-                await context.LibraryFolders.AddAsync(
+                await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+                await dbContext.LibraryFolders.AddAsync(
                     new LibraryFolder
                     {
                         Path = path,
                         Etag = etag,
                         LibraryPathId = libraryPath.Id
                     });
-                await context.SaveChangesAsync();
+                await dbContext.SaveChangesAsync();
             }).ToUnit();
+    }
 
     public async Task<Unit> CleanEtagsForLibraryPath(LibraryPath libraryPath)
     {
-        IEnumerable<string> folders = await _dbConnection.QueryAsync<string>(
+        await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync();
+
+        IEnumerable<string> folders = await dbContext.Connection.QueryAsync<string>(
             @"SELECT LF.Path
                 FROM LibraryFolder LF
                 WHERE LF.LibraryPathId = @LibraryPathId",
@@ -143,7 +159,7 @@ public class LibraryRepository : ILibraryRepository
 
         foreach (string folder in folders.Where(f => !_localFileSystem.FolderExists(f)))
         {
-            await _dbConnection.ExecuteAsync(
+            await dbContext.Connection.ExecuteAsync(
                 @"DELETE FROM LibraryFolder WHERE LibraryPathId = @LibraryPathId AND Path = @Path",
                 new { LibraryPathId = libraryPath.Id, Path = folder });
         }
